+++ packages/hooks/src/subAccount/useSubAccountMaxWithdrawal.ts
@@ -2,8 +2,8 @@ import { useMemo } from "react";
 import { account as accountPerp } from "@orderly.network/perp";
 import { API } from "@orderly.network/types";
 import { Decimal } from "@orderly.network/utils";
+import { useAppStore } from "../orderly/appStore";
 import { useIndexPricesStream } from "../orderly/useIndexPricesStream";
-import { useTokenInfo } from "../orderly/useTokensInfo/tokensInfo.store";
 
 const { maxWithdrawalUSDC, maxWithdrawalOtherCollateral, collateralRatio } =
   accountPerp;
@@ -20,7 +20,9 @@ export function useSubAccountMaxWithdrawal(options: {
 }) {
   const { token, unsettledPnL, freeCollateral, holdings } = options;
 
-  const tokenInfo = useTokenInfo(token!);
+  const tokenInfo = useAppStore((state) =>
+    state.tokensInfo?.find((item) => item.token === token),
+  );
 
   const { data: indexPrices } = useIndexPricesStream();