+++ packages/ui-transfer/src/components/yieldBearingReminder/index.ts
@@ -1,0 +1,8 @@
+export { YieldBearingReminder } from "./yieldBearingReminder";
+export type { YieldBearingReminderProps } from "./yieldBearingReminder";
+export { YusdReminder } from "./yusdReminder";
+export { DeUsdReminder } from "./deUsdReminder";
+export type { YusdReminderProps } from "./yusdReminder";
+export type { DeUsdReminderProps } from "./deUsdReminder";
+export { BaseReminder } from "./baseReminder";
+export type { BaseReminderProps } from "./baseReminder";