+++ packages/vaults/src/components/all-vaults/index.ts
@@ -1,2 +1,4 @@
 export * from "./all-vaults.widget";
 export * from "./all-vaults.desktop";
+export * from "./view-mode-toggle";
+export * from "./vaults-list";