+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @orderly.network/hooks
 
+## 2.8.2
+
+### Patch Changes
+
+- publish
+- 437c752: publish
+- Updated dependencies
+- Updated dependencies [437c752]
+  - @orderly.network/core@2.8.2
+  - @orderly.network/default-evm-adapter@2.8.2
+  - @orderly.network/default-solana-adapter@2.8.2
+  - @orderly.network/net@2.8.2
+  - @orderly.network/perp@4.8.2
+  - @orderly.network/types@2.8.2
+  - @orderly.network/utils@2.8.2
+  - @orderly.network/web3-provider-ethers@2.8.2
+
+## 2.8.2-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/core@2.8.2-alpha.0
+  - @orderly.network/default-evm-adapter@2.8.2-alpha.0
+  - @orderly.network/default-solana-adapter@2.8.2-alpha.0
+  - @orderly.network/net@2.8.2-alpha.0
+  - @orderly.network/perp@4.8.2-alpha.0
+  - @orderly.network/types@2.8.2-alpha.0
+  - @orderly.network/utils@2.8.2-alpha.0
+  - @orderly.network/web3-provider-ethers@2.8.2-alpha.0
+
 ## 2.8.1
 
 ### Patch Changes