+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,18 @@
 # tsconfig
 
+## 0.11.2
+
+### Patch Changes
+
+- publish
+- 437c752: publish
+
+## 0.11.2-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 0.11.1
 
 ### Patch Changes